+++ gatsby-browser.js
@@ -1,3 +1,3 @@
-import { wrapRootElement as wrap } from './src/root-wrapper';
+import { wrapRootElement as wrap } from './root-wrapper';
 
 export const wrapRootElement = wrap;
